Representation

Representation

Polyhedra can be described in 2 different ways.

H-representation: As the intersection of finitely many halfspaces given by its facets.

V-representation: As the convex hull of its vertices + the conic hull of its rays where '+' is the Minkowski sum.

In Polyhedra.jl, those representations are given the respective abstract types HRepresentation and VRepresentation which are themself subtypes of Representation.

H-representation

The fundamental element of an H-representation is the halfspace

Polyhedra.HalfSpace — Type.

struct HalfSpace{N, T, AT} <: HRepElement{N, T}
    a::AT
    β::T
end

An halfspace defined by the set of points $x$ such that $\langle a, x \rangle \le \beta$.

source

An H-representation can be created as the intersection of several halfspaces. For instance, the polytope

\[\begin{align*} x_1 + x_2 &\leq 1 \\ x_1 - x_2 &\leq 0 \\ x_1 & \geq 0. \end{align*}\]

can be created as follows:

HalfSpace([1, 1], 1) ∩ HalfSpace([1, -1], 0) ∩ HalfSpace([-1, 0], 0)

Even if HalfSpaces are enough to describe any polyhedron, it is sometimes important to represent the fact that the polyhedron is contained in an affine subspace. For instance, the simplex:

\[\begin{align*} x_1 + x_2 &= 1 \\ x_1 &\geq 0 \\ x_2 &\geq 0 \end{align*}\]

is 1-dimensional even if it is defined in a 2-dimensional space.

The fundamental element of an affine subspace is the hyperplane

Polyhedra.HyperPlane — Type.

struct HyperPlane{N, T, AT} <: HRepElement{N, T}
    a::AT
    β::T
end

An hyperplane defined by the set of points $x$ such that $\langle a, x \rangle = \beta$.

source

An affine subspace can be created as the intersection of several hyperplanes. For instance

HyperPlane([1, 1], 1) ∩ HyperPlane([1, 0], 0)

represents the 0-dimensional affine subspace only containing the point $(0, 1)$.

To represent a polyhedron that is not full-dimensional, hyperplanes and halfspaces can be mixed in any order. For instance, the simplex defined above can be obtained as follows:

HalfSpace([-1, 0], 0) ∩ HyperPlane([1, 1], 1) ∩ HalfSpace([0, -1], 0)

In addition to being created incrementally with intersections, an H-representation can also be created using the hrep function

Polyhedra.hrep — Function.

hrep(p::Polyhedron)

Returns an H-representation for the polyhedron p.

source

hrep(hyperplanes::ElemIt{<:HyperPlane})

Creates an affine space from the list of hyperplanes hyperplanes.

Examples

hrep([HyperPlane([0, 1, 0], 1), HyperPlane([0, 0, 1], 0)])

creates the 1-dimensional affine subspace containing all the points $(x_1, 0, 0)$, i.e. the x_1-axis.

hrep([HyperPlane([1, 1], 1), HyperPlane([1, 0], 0)])

creates the 0-dimensional affine subspace only containing the point $(0, 1)$.

source

hrep(A::AbstractMatrix, b::AbstractVector, linset::IntSet=IntSet())

Creates an H-representation for the polyhedron defined by the inequalities $\langle A_i, x \rangle = b_i$ if i in linset and $\langle A_i, x \rangle \le b_i$ otherwise where $A_i$ is the $i$th row of A, i.e. A[i,:] and $b_i$ is b[i].

source

Interface

An H-representation is represented as an intersection halfspaces and hyperplanes. The halfspaces can be obtained with halfspaces and the hyperplanes with hyperplane. As an hyperplane $\langle a, x \rangle = \beta$ is the intersection of the two halfspaces $\langle a, x \rangle \le \beta$ and $\langle a, x \rangle \ge \beta$, even if the H-representation contains hyperplanes, a list of halfspaces whose intersection is the polyhedron can be obtained with allhalfspaces, which has nhalfspaces(p) + 2nhyperplanes(p) elements for an H-representation p since each hyperplane is split in two halfspaces.

Polyhedra.halfspaces — Function.

halfspaces(hrep::HRep)

Returns an iterator over the halfspaces of the H-representation hrep.

source

Polyhedra.nhalfspaces — Function.

nhalfspaces(hrep::HRep)

Returns the number of halfspaces of the H-representation hrep.

source

Polyhedra.hashalfspaces — Function.

hashalfspaces(hrep::HRep)

Returns whether the H-representation hrep has any halfspace.

source

Polyhedra.hyperplanes — Function.

hyperplanes(hrep::HRep)

Returns an iterator over the hyperplanes of the H-representation hrep.

source

Polyhedra.nhyperplanes — Function.

nhyperplanes(hrep::HRep)

Returns the number of hyperplanes of the H-representation hrep.

source

Polyhedra.hashyperplanes — Function.

hashyperplanes(hrep::HRep)

Returns whether the H-representation hrep has any hyperplane.

source

Polyhedra.allhalfspaces — Function.

allhalfspaces(hrep::HRep)

Returns an iterator over the halfspaces and hyperplanes in the H-representation hrep splitting hyperplanes in two halfspaces.

Examples

hrep = HyperPlane([1, 0], 1) ∩ HalfSpace([0, 1], 1)
collect(allhalfspaces(hrep)) # Returns [HalfSpace([1, 0]), HalfSpace([-1, 0]), HalfSpace([0, 1])]

source

Polyhedra.nallhalfspaces — Function.

nallhalfspaces(hrep::HRep)

Returns the number of halfspaces plus twice the number of hyperplanes in the H-representation hrep, i.e. length(allhalfspaces(hrep))

source

Polyhedra.hasallhalfspaces — Function.

hasallhalfspaces(hrep::HRep)

Returns whether the H-representation hrep contains any halfspace or hyperplane.

source

Polyhedra.hrepiscomputed — Function.

hrepiscomputed(p::Polyhedron)

Returns whether the H-representation of this polyhedron has been computed.

source

V-representation

For instance, consider the 2-dimensional polyhedron described by the following H-representation:

\[\begin{align*} x_1 + x_2 &\leq 1 \\ x_1 - x_2 &\leq 0 \\ x_1 & \geq 0. \end{align*}\]

This set of inequalities can be written in the matrix form $Ax \leq b$ where

\[A = \begin{pmatrix}1 & 1\\1 & -1\\-1 & 0\end{pmatrix}, b = \begin{pmatrix}1\\0\\0\end{pmatrix}.\]

Let's create this H-representation using the concrete subtype SimpleHRepresentation of the abstract type HRepresentation.

julia> using Polyhedra
julia> A = [1 1;1 -1;-1 0]
julia> b = [1,0,0]
julia> hrep = SimpleHRepresentation(A, b)
julia> typeof(hrep)
Polyhedra.SimpleHRepresentation{2,Int64}

This polyhedron has three vertices: $(0,0)$, $(0,1)$ and $(0.5,0.5)$. We can create this V-representation using the concrete subtype SimpleVRepresentation of the abstract type VRepresentation. Because $0.5$ is fractional, have two choices: either use exact rational arithemtic

julia> V = [0 0; 0 1; 1//2 1//2]
julia> vrep = SimpleVRepresentation(V)
julia> typeof(vrep)
Polyhedra.SimpleVRepresentation{2,Rational{Int64}}

or use floating point arithmetic

julia> Vf = [0 0; 0 1; 1/2 1/2]
julia> vrepf = SimpleVRepresentation(Vf)
julia> typeof(vrepf)
Polyhedra.SimpleVRepresentation{2,Float64}

Representation interface

These functions can be called on both H-representation and V-representation

Polyhedra.fulldim — Function.

fulldim(rep::Rep)

Returns the dimension of the space in which the representation is defined. That is, a straight line in a 3D space has fulldim 3.

source

V-representation interface

Polyhedra.points — Function.

points(vrep::VRep)

Returns an iterator over the points of the V-representation vrep.

source

Polyhedra.npoints — Function.

npoints(vrep::VRep)

Returns the number of points of the V-representation vrep.

source

Polyhedra.haspoints — Function.

haspoints(vrep::VRep)

Returns whether the V-representation vrep has any point.

source

Polyhedra.sympoints — Function.

sympoints(vrep::VRep)

Returns an iterator over the sympoints of the V-representation vrep.

source

Polyhedra.nsympoints — Function.

nsympoints(vrep::VRep)

Returns the number of sympoints of the V-representation vrep.

source

Polyhedra.hassympoints — Function.

hassympoints(vrep::VRep)

Returns whether the V-representation vrep has any sympoint.

source

Polyhedra.allpoints — Function.

allpoints(vrep::VRep)

Returns an iterator over the points and sympoints in the V-representation vrep splitting sympoints in two points.

Examples

vrep = convexhull(SymPoint([1, 0]), [0, 1])
collect(allpoints(vrep)) # Returns [[1, 0], [-1, 0], [0, 1]]

source

Polyhedra.nallpoints — Function.

nallpoints(vrep::VRep)

Returns the number of points plus twice the number of sympoints in the V-representation vrep, i.e. length(allpoints(vrep))

source

Polyhedra.hasallpoints — Function.

hasallpoints(vrep::VRep)

Returns whether the V-representation vrep contains any point or sympoint.

source

Polyhedra.rays — Function.

rays(vrep::VRep)

Returns an iterator over the rays of the V-representation vrep.

source

Polyhedra.nrays — Function.

nrays(vrep::VRep)

Returns the number of rays of the V-representation vrep.

source

Polyhedra.hasrays — Function.

hasrays(vrep::VRep)

Returns whether the V-representation vrep has any ray.

source

Polyhedra.lines — Function.

lines(vrep::VRep)

Returns an iterator over the lines of the V-representation vrep.

source

Polyhedra.nlines — Function.

nlines(vrep::VRep)

Returns the number of lines of the V-representation vrep.

source

Polyhedra.haslines — Function.

haslines(vrep::VRep)

Returns whether the V-representation vrep has any line.

source

Polyhedra.allrays — Function.

allrays(vrep::VRep)

Returns an iterator over the rays and lines in the V-representation vrep splitting lines in two rays.

Examples

vrep = Line([1, 0]) + Ray([0, 1])
collect(allrays(vrep)) # Returns [Ray([1, 0]), Ray([-1, 0]), Ray([0, 1])]

source

Polyhedra.nallrays — Function.

nallrays(vrep::VRep)

Returns the number of rays plus twice the number of lines in the V-representation vrep, i.e. length(allrays(vrep))

source

Polyhedra.hasallrays — Function.

hasallrays(vrep::VRep)

Returns whether the V-representation vrep contains any ray or line.

source

Polyhedra.vrepiscomputed — Function.

vrepiscomputed(p::Polyhedron)

Returns whether the V-representation of this polyhedron has been computed.

source