Redundancy

Redundancy

Polyhedra.dim — Function.

dim(p::Polyhedron)

Returns the dimension of the affine hull of the polyhedron. That is the number of non-redundant hyperplanes that define it.

source

Polyhedra.isredundant — Function.

isredundant(p::Rep, idx::Index; strongly=false)

Return a Bool indicating whether the element with index idx can be removed without changing the polyhedron represented by p. If strongly is true,

if idx is an H-representation element h, it returns true only if no V-representation element of p is in the hyperplane of h.

if idx is a V-representation element v, it returns true only if v is in the relative interior of p.

source

Polyhedra.removehredundancy! — Function.

removehredundancy!(p::HRep)

Removes the elements of the H-representation of p that can be removed without changing the polyhedron represented by p. That is, it only keeps the halfspaces corresponding to facets of the polyhedron.

source

Polyhedra.removevredundancy! — Function.

removevredundancy!(p::VRep)

Removes the elements of the V-representation of p that can be removed without changing the polyhedron represented by p. That is, it only keeps the extreme points and rays. This operation is often called "convex hull" as the remaining points are the extreme points of the convex hull of the initial set of points.

source