Representation

Representation

Polyhedra can be described in 2 different ways.

H-representation: As the intersection of finitely many halfspaces given by its facets.

V-representation: As the convex hull of its vertices + the conic hull of its rays where '+' is the Minkowski sum.

In Polyhedra.jl, those representations are given the respective abstract types HRepresentation and VRepresentation which are themself subtypes of Representation.

For instance, consider the 2-dimensional polyhedron described by the following H-representation:

\[\begin{align*} x_1 + x_2 &\leq 1 \\ x_1 - x_2 &\leq 0 \\ x_1 & \geq 0. \end{align*}\]

This set of inequalities can be written in the matrix form $Ax \leq b$ where

\[A = \begin{pmatrix}1 & 1\\1 & -1\\-1 & 0\end{pmatrix}, b = \begin{pmatrix}1\\0\\0\end{pmatrix}.\]

Let's create this H-representation using the concrete subtype SimpleHRepresentation of the abstract type HRepresentation.

julia> using Polyhedra
julia> A = [1 1;1 -1;-1 0]
julia> b = [1,0,0]
julia> hrep = SimpleHRepresentation(A, b)
julia> typeof(hrep)
Polyhedra.SimpleHRepresentation{2,Int64}

This polyhedron has three vertices: $(0,0)$, $(0,1)$ and $(0.5,0.5)$. We can create this V-representation using the concrete subtype SimpleVRepresentation of the abstract type VRepresentation. Because $0.5$ is fractional, have two choices: either use exact rational arithemtic

julia> V = [0 0; 0 1; 1//2 1//2]
julia> vrep = SimpleVRepresentation(V)
julia> typeof(vrep)
Polyhedra.SimpleVRepresentation{2,Rational{Int64}}

or use floating point arithmetic

julia> Vf = [0 0; 0 1; 1/2 1/2]
julia> vrepf = SimpleVRepresentation(Vf)
julia> typeof(vrepf)
Polyhedra.SimpleVRepresentation{2,Float64}

Representation interface

These functions can be called on both H-representation and V-representation

Polyhedra.fulldim — Function.

fulldim(rep::Rep)

Returns the dimension of the space in which the representation is defined. That is, a straight line in a 3D space has fulldim 3.

source

H-representation interface

Polyhedra.halfspaces — Function.

halfspaces(hrep::HRep)

Returns an iterator over the halfspaces of the H-representation hrep.

source

Polyhedra.nhalfspaces — Function.

nhalfspaces(hrep::HRep)

Returns the number of halfspaces of the H-representation hrep.

source

Polyhedra.hashalfspaces — Function.

hashalfspaces(hrep::HRep)

Returns whether the H-representation hrep has any halfspace.

source

Polyhedra.hyperplanes — Function.

hyperplanes(hrep::HRep)

Returns an iterator over the hyperplanes of the H-representation hrep.

source

Polyhedra.nhyperplanes — Function.

nhyperplanes(hrep::HRep)

Returns the number of hyperplanes of the H-representation hrep.

source

Polyhedra.hashyperplanes — Function.

hashyperplanes(hrep::HRep)

Returns whether the H-representation hrep has any hyperplane.

source

Polyhedra.allhalfspaces — Function.

allhalfspaces(hrep::HRep)

Returns an iterator over the halfspaces and hyperplanes in the H-representation hrep splitting hyperplanes in two halfspaces.

Examples

hrep = HyperPlane([1, 0], 1) ∩ HalfSpace([0, 1], 1)
collect(allhalfspaces(hrep)) # Returns [HalfSpace([1, 0]), HalfSpace([-1, 0]), HalfSpace([0, 1])]

source

Polyhedra.nallhalfspaces — Function.

nallhalfspaces(hrep::HRep)

Returns the number of halfspaces plus twice the number of hyperplanes in the H-representation hrep, i.e. length(allhalfspaces(hrep))

source

Polyhedra.hasallhalfspaces — Function.

hasallhalfspaces(hrep::HRep)

Returns whether the H-representation hrep contains any halfspace or hyperplane.

source

Polyhedra.hrepiscomputed — Function.

hrepiscomputed(p::Polyhedron)

Returns whether the H-representation of this polyhedron has been computed.

source

V-representation interface

Polyhedra.points — Function.

points(vrep::VRep)

Returns an iterator over the points of the V-representation vrep.

source

Polyhedra.npoints — Function.

npoints(vrep::VRep)

Returns the number of points of the V-representation vrep.

source

Polyhedra.haspoints — Function.

haspoints(vrep::VRep)

Returns whether the V-representation vrep has any point.

source

Polyhedra.sympoints — Function.

sympoints(vrep::VRep)

Returns an iterator over the sympoints of the V-representation vrep.

source

Polyhedra.nsympoints — Function.

nsympoints(vrep::VRep)

Returns the number of sympoints of the V-representation vrep.

source

Polyhedra.hassympoints — Function.

hassympoints(vrep::VRep)

Returns whether the V-representation vrep has any sympoint.

source

Polyhedra.allpoints — Function.

allpoints(vrep::VRep)

Returns an iterator over the points and sympoints in the V-representation vrep splitting sympoints in two points.

Examples

vrep = convexhull(SymPoint([1, 0]), [0, 1])
collect(allpoints(vrep)) # Returns [[1, 0], [-1, 0], [0, 1]]

source

Polyhedra.nallpoints — Function.

nallpoints(vrep::VRep)

Returns the number of points plus twice the number of sympoints in the V-representation vrep, i.e. length(allpoints(vrep))

source

Polyhedra.hasallpoints — Function.

hasallpoints(vrep::VRep)

Returns whether the V-representation vrep contains any point or sympoint.

source

Polyhedra.rays — Function.

rays(vrep::VRep)

Returns an iterator over the rays of the V-representation vrep.

source

Polyhedra.nrays — Function.

nrays(vrep::VRep)

Returns the number of rays of the V-representation vrep.

source

Polyhedra.hasrays — Function.

hasrays(vrep::VRep)

Returns whether the V-representation vrep has any ray.

source

Polyhedra.lines — Function.

lines(vrep::VRep)

Returns an iterator over the lines of the V-representation vrep.

source

Polyhedra.nlines — Function.

nlines(vrep::VRep)

Returns the number of lines of the V-representation vrep.

source

Polyhedra.haslines — Function.

haslines(vrep::VRep)

Returns whether the V-representation vrep has any line.

source

Polyhedra.allrays — Function.

allrays(vrep::VRep)

Returns an iterator over the rays and lines in the V-representation vrep splitting lines in two rays.

Examples

vrep = Line([1, 0]) + Ray([0, 1])
collect(allrays(vrep)) # Returns [Ray([1, 0]), Ray([-1, 0]), Ray([0, 1])]

source

Polyhedra.nallrays — Function.

nallrays(vrep::VRep)

Returns the number of rays plus twice the number of lines in the V-representation vrep, i.e. length(allrays(vrep))

source

Polyhedra.hasallrays — Function.

hasallrays(vrep::VRep)

Returns whether the V-representation vrep contains any ray or line.

source

Polyhedra.vrepiscomputed — Function.

vrepiscomputed(p::Polyhedron)

Returns whether the V-representation of this polyhedron has been computed.

source